【修正・追記】モーダル機能実装（UIはまだデフォルト）
