+++ components/ui/DiaryDialog/index.tsx
@@ -6,22 +6,14 @@ import DialogContent from "@material-ui/core/DialogContent";
 import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
 
-const DiaryDialog = () => {
-  const [open, setOpen] = React.useState(false);
+interface Props {
+  open: boolean;
+  handleClose: () => void;
+}
 
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
+const DiaryDialog: React.FC<Props> = ({ open, handleClose }) => {
   return (
     <div>
-      <Button variant="outlined" color="primary" onClick={handleClickOpen}>
-        Open alert dialog
-      </Button>
       <Dialog
         open={open}
         onClose={handleClose}
